move files into new arranged positions
