+++ src/SmartVectorDotNet/ScalarOp/ScalarOp._BitOperations.BitHash.cs
