+++ src/SmartVectorDotNet/ScalarOp/ScalarOp._BitOperations.CountLeadingZeros.cs
