+++ src/SmartVectorDotNet/ScalarOp/ScalarOp._BitOperations.CountTrailingZeros.cs
