+++ src/SmartVectorDotNet/ScalarOp/ScalarOp._BitOperations.IEEE754Specific.cs
