+++ src/SmartVectorDotNet/ScalarOp/ScalarOp._Const.tt
