+++ src/SmartVectorDotNet/ScalarOp/ScalarOp._Convert.tt
