+++ src/SmartVectorDotNet/ScalarOp/ScalarOp._Math.Extensions.cs
