+++ src/SmartVectorDotNet/ScalarOp/ScalarOp._Math.General.tt
