+++ src/SmartVectorDotNet/ScalarOp/ScalarOp._Math.Real.tt
