+++ src/SmartVectorDotNet/ScalarOp/ScalarOp._Math.netstandard20.cs
