+++ src/SmartVectorDotNet/ScalarOp/ScalarOp._Operator.Binary.tt
