+++ src/SmartVectorDotNet/ScalarOp/ScalarOp._Operator.Comparison.cs
