+++ src/SmartVectorDotNet/ScalarOp/ScalarOp._Operator.Shift.tt
