+++ src/SmartVectorDotNet/ScalarOp/ScalarOp._Operator.Unary.tt
