+++ src/SmartVectorDotNet/ScalarOp/ScalarOp._OperatorEx.Saturate.cs
