+++ src/SmartVectorDotNet/SmartVectorDotNet.csproj
@@ -30,99 +30,99 @@
     </ItemGroup>
 
     <!-- #region ScalarOp -->
-    <ItemGroup Label="ScalarOp\ScalarOp._Binary">
-        <None Update="ScalarOp\ScalarOp._Binary.tt">
+    <ItemGroup Label="ScalarOp\ScalarOp._Convert">
+        <None Update="ScalarOp\ScalarOp._Convert.tt">
             <Generator>TextTemplatingFileGenerator</Generator>
-            <LastGenOutput>ScalarOp._Binary.g.cs</LastGenOutput>
+            <LastGenOutput>ScalarOp._Convert.g.cs</LastGenOutput>
         </None>
-        <Compile Update="ScalarOp\ScalarOp._Binary.g.cs">
+        <Compile Update="ScalarOp\ScalarOp._Convert.g.cs">
             <DesignTime>True</DesignTime>
             <AutoGen>True</AutoGen>
-            <DependentUpon>ScalarOp._Binary.tt</DependentUpon>
-        </Compile>
-    </ItemGroup>
-
-    <ItemGroup Label="ScalarOp\ScalarOp._Cast">
-        <None Update="ScalarOp\ScalarOp._Cast.tt">
-            <Generator>TextTemplatingFileGenerator</Generator>
-            <LastGenOutput>ScalarOp._Cast.g.cs</LastGenOutput>
-        </None>
-        <Compile Update="ScalarOp\ScalarOp._Cast.g.cs">
-            <DesignTime>True</DesignTime>
-            <AutoGen>True</AutoGen>
-            <DependentUpon>ScalarOp._Cast.tt</DependentUpon>
-        </Compile>
-    </ItemGroup>
-
-    <ItemGroup Label="ScalarOp\ScalarOp._Comparison">
-        <None Update="ScalarOp\ScalarOp._Comparison.tt">
-            <Generator>TextTemplatingFileGenerator</Generator>
-            <LastGenOutput>ScalarOp._Comparison.g.cs</LastGenOutput>
-        </None>
-        <Compile Update="ScalarOp\ScalarOp._Comparison.g.cs">
-            <DesignTime>True</DesignTime>
-            <AutoGen>True</AutoGen>
-            <DependentUpon>ScalarOp._Comparison.tt</DependentUpon>
+            <DependentUpon>ScalarOp._Convert.tt</DependentUpon>
         </Compile>
     </ItemGroup>
 
     <ItemGroup Label="ScalarOp\ScalarOp._Constant">
-        <None Update="ScalarOp\ScalarOp._Constant.tt">
+        <None Update="ScalarOp\ScalarOp._Const.tt">
             <Generator>TextTemplatingFileGenerator</Generator>
-            <LastGenOutput>ScalarOp._Constant.g.cs</LastGenOutput>
+            <LastGenOutput>ScalarOp._Const.g.cs</LastGenOutput>
         </None>
-        <Compile Update="ScalarOp\ScalarOp._Constant.g.cs">
+        <Compile Update="ScalarOp\ScalarOp._Const.g.cs">
             <DesignTime>True</DesignTime>
             <AutoGen>True</AutoGen>
-            <DependentUpon>ScalarOp._Constant.tt</DependentUpon>
+            <DependentUpon>ScalarOp._Const.tt</DependentUpon>
         </Compile>
     </ItemGroup>
 
-    <ItemGroup Label="ScalarOp\ScalarOp._General">
-        <None Update="ScalarOp\ScalarOp._General.tt">
+    <ItemGroup Label="ScalarOp\ScalarOp._Math.General">
+        <None Update="ScalarOp\ScalarOp._Math.General.tt">
             <Generator>TextTemplatingFileGenerator</Generator>
-            <LastGenOutput>ScalarOp._General.g.cs</LastGenOutput>
+            <LastGenOutput>ScalarOp._Math.General.g.cs</LastGenOutput>
         </None>
-        <Compile Update="ScalarOp\ScalarOp._General.g.cs">
+        <Compile Update="ScalarOp\ScalarOp._Math.General.g.cs">
             <DesignTime>True</DesignTime>
             <AutoGen>True</AutoGen>
-            <DependentUpon>ScalarOp._General.tt</DependentUpon>
+            <DependentUpon>ScalarOp._Math.General.tt</DependentUpon>
         </Compile>
     </ItemGroup>
 
-    <ItemGroup Label="ScalarOp\ScalarOp._Real">
-        <None Update="ScalarOp\ScalarOp._Real.tt">
+    <ItemGroup Label="ScalarOp\ScalarOp._Math.Real">
+        <None Update="ScalarOp\ScalarOp._Math.Real.tt">
             <Generator>TextTemplatingFileGenerator</Generator>
-            <LastGenOutput>ScalarOp._Real.g.cs</LastGenOutput>
+            <LastGenOutput>ScalarOp._Math.Real.g.cs</LastGenOutput>
         </None>
-        <Compile Update="ScalarOp\ScalarOp._Real.g.cs">
+        <Compile Update="ScalarOp\ScalarOp._Math.Real.g.cs">
             <DesignTime>True</DesignTime>
             <AutoGen>True</AutoGen>
-            <DependentUpon>ScalarOp._Real.tt</DependentUpon>
+            <DependentUpon>ScalarOp._Math.Real.tt</DependentUpon>
         </Compile>
     </ItemGroup>
 
-    <ItemGroup Label="ScalarOp\ScalarOp._Shift">
-        <None Update="ScalarOp\ScalarOp._Shift.tt">
+    <ItemGroup Label="ScalarOp\ScalarOp._Operator.Binary">
+        <None Update="ScalarOp\ScalarOp._Operator.Binary.tt">
             <Generator>TextTemplatingFileGenerator</Generator>
-            <LastGenOutput>ScalarOp._Shift.g.cs</LastGenOutput>
+            <LastGenOutput>ScalarOp._Operator.Binary.g.cs</LastGenOutput>
         </None>
-        <Compile Update="ScalarOp\ScalarOp._Shift.g.cs">
+        <Compile Update="ScalarOp\ScalarOp._Operator.Binary.g.cs">
             <DesignTime>True</DesignTime>
             <AutoGen>True</AutoGen>
-            <DependentUpon>ScalarOp._Shift.tt</DependentUpon>
+            <DependentUpon>ScalarOp._Operator.Binary.tt</DependentUpon>
         </Compile>
     </ItemGroup>
 
-    <ItemGroup Label="ScalarOp\ScalarOp._Unary">
-        <None Update="ScalarOp\ScalarOp._Unary.tt">
+    <ItemGroup Label="ScalarOp\ScalarOp._Operator.Shift">
+        <None Update="ScalarOp\ScalarOp._Operator.Shift.tt">
             <Generator>TextTemplatingFileGenerator</Generator>
-            <LastGenOutput>ScalarOp._Unary.g.cs</LastGenOutput>
+            <LastGenOutput>ScalarOp._Operator.Shift.g.cs</LastGenOutput>
         </None>
-        <Compile Update="ScalarOp\ScalarOp._Unary.g.cs">
+        <Compile Update="ScalarOp\ScalarOp._Operator.Shift.g.cs">
             <DesignTime>True</DesignTime>
             <AutoGen>True</AutoGen>
-            <DependentUpon>ScalarOp._Unary.tt</DependentUpon>
+            <DependentUpon>ScalarOp._Operator.Shift.tt</DependentUpon>
+        </Compile>
+    </ItemGroup>
+
+    <ItemGroup Label="ScalarOp\ScalarOp._Operator.Comparison">
+        <None Update="ScalarOp\ScalarOp._Operator.Comparison.tt">
+            <Generator>TextTemplatingFileGenerator</Generator>
+            <LastGenOutput>ScalarOp._Operator.Comparison.g.cs</LastGenOutput>
+        </None>
+        <Compile Update="ScalarOp\ScalarOp._Operator.Comparison.g.cs">
+            <DesignTime>True</DesignTime>
+            <AutoGen>True</AutoGen>
+            <DependentUpon>ScalarOp._Operator.Comparison.tt</DependentUpon>
+        </Compile>
+    </ItemGroup>
+
+    <ItemGroup Label="ScalarOp\ScalarOp._Operator.Unary">
+        <None Update="ScalarOp\ScalarOp._Operator.Unary.tt">
+            <Generator>TextTemplatingFileGenerator</Generator>
+            <LastGenOutput>ScalarOp._Operator.Unary.g.cs</LastGenOutput>
+        </None>
+        <Compile Update="ScalarOp\ScalarOp._Operator.Unary.g.cs">
+            <DesignTime>True</DesignTime>
+            <AutoGen>True</AutoGen>
+            <DependentUpon>ScalarOp._Operator.Unary.tt</DependentUpon>
         </Compile>
     </ItemGroup>
     <!-- #endregion ScalarOp -->
@@ -141,14 +141,14 @@
     </ItemGroup>
 
     <ItemGroup Label="VectorOp\VectorOp._Shift">
-        <Compile Update="VectorOp\VectorOp._Shift.g.cs">
+        <Compile Update="VectorOp\VectorOp._Operator.Shift.g.cs">
             <DesignTime>True</DesignTime>
             <AutoGen>True</AutoGen>
-            <DependentUpon>VectorOp._Shift.tt</DependentUpon>
+            <DependentUpon>VectorOp._Operator.Shift.tt</DependentUpon>
         </Compile>
-        <None Update="VectorOp\VectorOp._Shift.tt">
+        <None Update="VectorOp\VectorOp._Operator.Shift.tt">
             <Generator>TextTemplatingFileGenerator</Generator>
-            <LastGenOutput>VectorOp._Shift.g.cs</LastGenOutput>
+            <LastGenOutput>VectorOp._Operator.Shift.g.cs</LastGenOutput>
         </None>
     </ItemGroup>
