+++ src/SmartVectorDotNet/VectorOp/VectorOp._BitOperations.CountLeadingZeros.cs
