+++ src/SmartVectorDotNet/VectorOp/VectorOp._BitOperations.CountTrailingZeros.cs
