+++ src/SmartVectorDotNet/VectorOp/VectorOp._BitOperations.IEEE754Specific.cs
