+++ src/SmartVectorDotNet/VectorOp/VectorOp._MathEx.DivideLike.cs
