+++ src/SmartVectorDotNet/VectorOp/VectorOp._MathEx.ModuloByConst.cs
