+++ src/SmartVectorDotNet/VectorOp/VectorOp._Operator.Modulo.cs
