+++ src/SmartVectorDotNet/VectorOp/VectorOp._Operator.Shift.tt
