+++ src/SmartVectorDotNet/VectorOp/VectorOp._OperatorEx.Saturate.cs
