+++ src/SmartVectorDotNet/VectorOp/VectorOp._Vector.Permute2.cs
