+++ src/SmartVectorDotNet/VectorOp/VectorOp._Vector.Permute4.cs
