+++ src/SmartVectorDotNet/VectorOp/VectorOp._Vector.Permute_Internal.cs
